Small bug fixes. Moved classes to IBTK.
