+++ src/CMakeLists.txt
@@ -130,8 +130,6 @@ SET(CXX_SRC
   navier_stokes/INSCollocatedCenteredConvectiveOperator.cpp
   navier_stokes/StaggeredStokesSolver.cpp
   navier_stokes/INSVCStaggeredHierarchyIntegrator.cpp
-  navier_stokes/SnapshotCache.cpp
-  navier_stokes/HierarchyTimeInterpolator.cpp
 
   # IB
   IB/IBInstrumentPanel.cpp
